.history/requirements_20251031120125.txt
# requirements.txt
google-adk
pydantic
python-docx
reportlab
markdown
fastapi
uvicorn
python-multipart